Move helpers into separate file
